requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
requests
python-multipart
assemblyai
